Extend some client requests to be able to operate on multiple resources #1020 + Remove `info` in `GroupVersion` and add `lastUpdatedDate` in `Group` to support querying the last updated date of group on the client side efficiently + Trivial code improvements

+++ turms-client-swift/Tests/TurmsClientTests/Helper/Config.swift
@@ -1,5 +1,5 @@
 enum Config {
     static let HOST = "127.0.0.1"
-    static let PORT = 11510
+    static let PORT = UInt16(11510)
     static let STORAGE_SERVER_URL = "http://localhost:9000"
 }
